migrate database outside of the app itself
This changes the app so that it doesn't migrate its own database on app
startup.  This means that whenever the app runs, something else needs to
ensure the database has been migrated appropriately.  There are four
situations:

 - for actual deployment, the start-service.sh script runs the
   migration (using a command I manually tested on an existing EC2 box)
 - for ./run-application, we have an extra gradle task migrateMainDb
 - for functional tests, RegisterRule migrates the database before
   running the tests
 - for conformance tests, there was already the appropriate usage of
   {create,drop}_schema.sh, but the waitForApplicationToStart()
   operation broke because the app didn't return a 200 response on the
   application port when the database hadn't been migrated already.  I
   replaced this with a call to /healthcheck on the admin port instead.

The longer-term goal is to entirely separate application deployment from
database deployment, so that we can have automatic rollbacks for apps
and have a separate deployment pipeline for database migrations.